Updates README for API key usage; add sample config
Since explaining how to use an API key requires an understanding of how
the OAuth configuration works, this commit includes a brief outline
of how the `config.json` works as well as a sample. The verbiage in
the README was drawn from similar language in the
[singer-io/target-csv](https://github.com/singer-io/target-csv/blob/2e76485b2d574594387a945aeb7d6e7df4c43645/README.md)
repo.

Partially addresses issue #9.