Syncs contacts based on `versionTimestamp`
This commit addresses a long-standing issue with list memberships in
which memberships are dependent on whether or not a contact has been
updated, and thus new dynamic list memberships are frequently not
captured.

HubSpot has recently exposed a new pair of fields (`version` &
`versionTimestamp`) which are returned when the `includeVersion` query
parameter is set to `true` when calling the `contacts` APIs. These new
fields track the internal versions of the contacts, which includes
updates due to new dynamic list memberships. Updates to other contact
properties will also trigger a new `versionTimestamp`, so no prior
functionality is lost.