Use Deal v3 API
In order to properly get the deal stage entry / exit date,
we need to use the V3 Deals API.

Previous attempts were mixing the v1 and v3 APIs but weren't
functional.

This approach use only v3 API. However, we've lost the "versions"
of the properties that the v1 API was providing.